Kiev21 devs 261 fixbugs (#12)
* KIEV21DEVS-123 Configure checkstyle modules and refactoring.

* KIEV21DEVS-261 Fix bugs and refactoring.

* Test

* KIEV21DEVS-261 Fix bugs and refactoring.

+++ src/main/java/com/heroku/spacey/dao/MaterialDao.java
@@ -3,11 +3,11 @@ package com.heroku.spacey.dao;
 import com.heroku.spacey.entity.Material;
 
 public interface MaterialDao {
-    Material getById(int id);
+    Material getById(Long id);
 
-    int insert(Material material);
+    Long insert(Material material);
 
     void update(Material material);
 
-    void delete(int id);
+    void delete(Long id);
 }
