+++ src/main/java/com/heroku/spacey/services/ProductService.java
@@ -5,13 +5,13 @@ import com.heroku.spacey.dto.product.ProductDto;
 import com.heroku.spacey.dto.product.UpdateProductDto;
 
 public interface ProductService {
-    ProductDto getById(int id);
+    ProductDto getById(Long id);
 
     void addProduct(AddProductDto addProductDto);
 
-    void updateProduct(UpdateProductDto updateProductDto, int id);
+    void updateProduct(UpdateProductDto updateProductDto, Long id);
 
-    void removeProduct(int id);
+    void removeProduct(Long id);
 
-    void cancelProduct(int id);
+    void cancelProduct(Long id);
 }
